Added source param for DataStream name and S3 folder name
post.request.data:
{
"line": "log line here",
"source": "Android"
}
For each "source", aws will create a delivery stream and use it as the
folder name in S3.
So better to limit the number of "source"!

Limit source in ['Android','Ibetweb']?

If no "source" is provided, an UnknownDeliveryStream will be created and
S3 folder name would be "Unknown/"